astu-store-dev-1.0/requirements/base.txt
dj-database-url==0.5.0
Django==3.2.4
psycopg2-binary==2.9.1
python-decouple==3.6
pytz==2022.1
Pillow==9.1.1
django-widget-tweaks==1.4.11
django-smart-selects==1.5.9
django-summernote==0.8.20.0
